Edited message handler
Now will see if a match begin message addresses our player, if so he clears the list of past moves for his new game.
Also changed match_end to see if the message is addressing us before adding the move to the list. Before the player would add any move that wasn't his own to the list. Now he ensures that he is one of the players in the match, and then adds the corresponding move appropriately.